Refactor mail subscription commands and handlers
Update import paths and remove deprecated files
