+++ src/auth/repositories/token.repository.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
+import { Token } from 'src/auth/entities/token.schema';
 import { ITokenDocument } from 'src/auth/interfaces/token.interface';
 import { GenericRepository } from 'src/common/providers/generic.repository';
-import { Token } from '../entities/token.schema';
 
 @Injectable()
 export class TokenRepository extends GenericRepository<ITokenDocument> {
