+++ src/bucket/entities/bucket.schema.ts
@@ -1,6 +1,6 @@
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 import { UploadApiResponse } from 'cloudinary';
-import { IBucket } from '../interfaces/bucket.interface';
+import { IBucket } from 'src/bucket/interfaces/bucket.interface';
 
 @Schema({
   timestamps: true,
