+++ src/bucket/repositories/bucket.repository.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
+import { Bucket } from 'src/bucket/entities/bucket.schema';
+import { IBucketDocument } from 'src/bucket/interfaces/bucket.interface';
 import { GenericRepository } from 'src/common/providers/generic.repository';
-import { Bucket } from '../entities/bucket.schema';
-import { IBucketDocument } from '../interfaces/bucket.interface';
 
 @Injectable()
 export class BucketRepository extends GenericRepository<IBucketDocument> {
