+++ src/mail/commands/remove-subscription/remove-subscription.command.ts
@@ -1,0 +1,7 @@
+import { Command } from '@nestjs/cqrs';
+
+export class RemoveMailSubscriptionCommand extends Command<void> {
+  constructor(public readonly id: string) {
+    super();
+  }
+}
