+++ src/mail/controllers/mail.controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { MailService } from '../services/mail.service';
+import { MailService } from 'src/mail/services/mail.service';
 import { MailController } from './mail.controller';
 
 describe('MailController', () => {
