+++ src/mail/mail.module.ts
@@ -2,7 +2,8 @@ import { Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
 import { MongooseModule } from '@nestjs/mongoose';
 import { BroadcastMailHandler } from './commands/broadcast-mail/broadcast-mail.handler';
-import { CreateMailSubscriptionHandler } from './commands/create-mail-subscription/create-mail-subscription.handler';
+import { CreateMailSubscriptionHandler } from './commands/create-subscription/create-subscription.handler';
+import { RemoveMailSubscriptionHandler } from './commands/remove-subscription/remove-subscription.handler';
 import { SendSimpleMailHandler } from './commands/send-mail/send-mail.handler';
 import { MailController } from './controllers/mail.controller';
 import { MailSubscriptionController } from './controllers/subscription.controller';
@@ -31,6 +32,7 @@ import { MailService } from './services/mail.service';
     BroadcastMailHandler,
     FindAllMailSubscriptionsHandler,
     FindOneMailSubscriptionHandler,
+    RemoveMailSubscriptionHandler,
   ],
   exports: [MailService],
 })
