+++ src/mail/repositories/subscription.repository.ts
@@ -2,8 +2,8 @@ import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { GenericRepository } from 'src/common/providers/generic.repository';
-import { MailSubscription } from '../entities/subscription.schema';
-import { IMailSubscriptionDocument } from '../interfaces/subscription.interface';
+import { MailSubscription } from 'src/mail/entities/subscription.schema';
+import { IMailSubscriptionDocument } from 'src/mail/interfaces/subscription.interface';
 
 @Injectable()
 export class MailSubscriptionRepository extends GenericRepository<IMailSubscriptionDocument> {
