+++ src/user/repositories/user.repository.ts
@@ -3,7 +3,7 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { GenericRepository } from 'src/common/providers/generic.repository';
 import { User } from 'src/user/entities/user.schema';
-import { IUserDocument } from '../interfaces/user.interface';
+import { IUserDocument } from 'src/user/interfaces/user.interface';
 
 @Injectable()
 export class UserRepository extends GenericRepository<IUserDocument> {
