+++ test/app.e2e-spec.ts
@@ -1,7 +1,7 @@
 import { INestApplication } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
-import * as request from 'supertest';
-import { AppModule } from './../src/app.module';
+import { AppModule } from 'src/app.module';
+import request from 'supertest';
 
 describe('AppController (e2e)', () => {
   let app: INestApplication;
